requirements.txt
piexif
PyYAML